Remove the password element from the generated dict in update_user
This prevents a log warning when a hashed password is too long for the
configured password_hashing_algorithm, as the currently hashed password
is passed into the from_dict method, which will try to 'set' the password
using the hashed value.

Also, handling of a password update is done later in this method, so the
warning message would also be misleading anyone interpreting the log lines,
as the password is not updated using this logic.

Closes-Bug: #2060441
Change-Id: I53d66fe87179aac7b3e211c84fd9858b1157b352