Improve configuration of out-of-tree identity drivers
Official recommendation for operators is to develop out-of-tree drivers
for the missing functionality. While this is easily possible there are
hurdles for configuring such drivers. Configuring them using FS requires
restart of Keystone when i.e. adding new domains to be processed with
the custom driver. Using database is a much better and dynamical
approach, but it currently is not allowing drivers to have specific
configuration. This change improves this flaws and contains of 3
individual parts that are submitted together to make testing easier.

- Allowing driver to register supported configuration option before
  loading the driver (invoke DriverManager without calling the driver
  and search for specific method present).
- Allow changing driver specific configuration through API (add 2 more
  configuration options enabling listed options in the API)
- Documentation changes.

Change-Id: I99fa798ef60cdb7a488fe55de76cd931c6db3e89