Enable non-voting OpenAPI build job
We are going to start building OpenAPI using codegenerator directly in
the Keystone and move all the schemas into the codebasis. It is going to
be very helpful to immediately see whether generation is possible and
the OpenAPI spec is valid. Thus adding it as non-voting for now (and it
makes no sense of having non-voting job in gate).

Change-Id: Iaf418d69cbb55fbe9dde3f2ec4213e64345057c2