Add keystone-manage reset_last_active command
This patch adds the `reset_last_active` subcommand to the
`keystone-manage` command line tool.

This subcommand will update every user in the database that has a null
value in the `last_active_at` property to the current server time. This
is necessary to prevent user lockout in deployments that have been
running for a long time without `disable_user_account_days_inactive` and
later decide to turn it on.

This patch also includes a change to the logic that sets
`last_active_at` to fix the root issue of the lockout.

Closes-Bug: 2074018
Change-Id: I1b71fb3881dc041db01083fbb4f2592400096a31
(cherry picked from commit e9513f8e4f25e1f20bc6fcab71d9177120000abf)