extend docs explaining OIDC
OpenID Connect has specific terms for different parts so add those to
the documentation or use them when the wrong term was used. Added some
more expanded details around the different settings and how they work.
mod_auth_openidc takes lists in claims and when it converts them to a
header it uses a comma as its delimiter by default. But keystone uses ;
as it's delimiter when using a field in a mapping as a list. So the user
must configure mod_auth_openidc to use a ; instead of a ,

Change-Id: I3c44ed0937679f01d554b2fa74c9916d54c68464