Switch to crypt from passlib
Next step in getting rid of `passlib` is sha512_crypt. This one is
tricky since the `crypt` lib is deprecated in py311 and dropped in py313
with the only replacement by `passlib`. But since passlib itself is
abandoned it literally means there is nothing we could use.

On the other hand `sha512_crypt` was "deprecated" already in pike and it
is not possible to enable it in Keystone while old passwords are still
supported.

In this unlucky situation still get rid of passlib and support for this
algorightm is going to be completely dropped in next Keystone release.

Change-Id: If791c953bc2953f2c78ff91e4fc4a342b89d6d1b