Fix DB migrations after alembic integration
With a change to swap sqlalchemy-migrate with alembic [1]
a `db_sync --check` was broken.

This is due to both `upgrades.get_db_version` and
`upgrades.get_current_heads` are actually checking "current" state of
the dabase by calling _get_current_heads[2][3], while obvious intention
was to compare intended state with current state.

With that we're introducing upgrade.get_head_revisions which will fetch
revisions not from the database, but from the environment [4]

As a result `db_sync --check` does compare desired state of the DB with
actual state and exists with corresponsive status again.

[1] https://opendev.org/openstack/keystone/commit/f174b4fa7c4fb010bbacc8c5a5f3625a8fcb41f3
[2] https://opendev.org/openstack/keystone/src/branch/master/keystone/common/sql/upgrades.py#L147
[3] https://opendev.org/openstack/keystone/src/branch/master/keystone/common/sql/upgrades.py#L191
[4] https://alembic.sqlalchemy.org/en/latest/api/runtime.html#alembic.runtime.environment.EnvironmentContext.get_head_revisions

Closes-Bug: #2080542
Change-Id: I854d37e3b4a34a7880f157564466bde61a3f886a