[stable-only] Cap setuptools <71.0.0 and fix gate
py39 jobs (on ubuntu-focal) started to fail due to recent virtualenv
release (20.26.4) on Yoga (which bundles setuptools), because we have
'packaging==21.3' in this branch that is not compatible with newer
setuptools [1].

setuptools is bundled in virtualenv, so it has to be capped via the
virtualenv package. tox also needed to be capped (<4) as gate uses
tox 3.28.0, but with capping virtualenv we pull in latest tox as well,
which would cause other errors.

This patch also sets to use keystone-tempest-plugin from wallaby-last
tag for functional jobs otherwise they fail [2] as on newer versions
of keystone-tempest-plugin keystoneauth is set as >=5.1.1 in
requirements.txt however we pin keystoneauth in upper-constraints.txt
much lower version than that.

[1] https://github.com/pypa/setuptools/issues/4483
[2] ERROR: Cannot install keystone-tempest-plugin==0.17.0 because these
package versions have conflicting dependencies.
The conflict is caused by:
    keystone-tempest-plugin 0.17.0 depends on keystoneauth1>=5.1.1
    The user requested (constraint) keystoneauth1===4.5.0

Change-Id: I76c39af6119806631d800ad644db270dd7d01b64