Unify response code for EC2_S3_Resource
When looking up the credential we dont catch the
CredentialNotFound exception which causes us to
return a 404 Not Found response code to the client
which is fine, but since we dont catch it we also
output the entire exception to stderr.

This changes so that we instead catch the
CredentialNotFound exception and raise a
NotFound exception which retains the behaviour
but does not pollute the logs with long tracebacks.

This also addresses a security concern if the
credential is found but is of the wrong type we
throw an Unauthorized response code but say in
the message that the "EC2 access key is not found",
this could potentially be used for doing a enumeration
attack trying to figure out if a credential exists
or not.

To fix this we change the response code from
Unauthorized to Not Found which makes it impossible
to know which part of the code raised the error
from the outside.

Closes-Bug: #2077541
Change-Id: I0ffeaf032ccdfcd99da27719cf90451a5855af81