Replace sphinxcontrib-*diag
As expected (and as done in lot of other OpenStack services) *blockdiag
and *seqdiag extensions stop working for us and we need to replace this
usage. Other services go directly to graphviz. Sadly sequence diagrams
doesn't look good, but there is no other alternative so far. It would be
nice if we would have mermaid allowed in global requirements, but that
tends to be a long process while we need an urgent solution (since
otherwise doc build is broken).

Change-Id: Ie5fc80d7e5506c8c426a289687acc6c7c30f5ebd