Fix wsgi service name for the federation tests
in devstack the service name was set explicitly to keystone-api as what
we have used for the new wsgi module. By default when unset devstack was
defaulting to the basename of the wsgi script (keystone-wsgi-public).
That file is created by federation tests but is not enabled in devstack.
Fix that.

Change-Id: Ie1b0e1e68441785e3b00543de75575c49917f591