Fix domain `immutable` lockdown
A bug in the test design have hidden the issue with options of the
domain. It is possible to set immutable to true during creation or
update of the domain, but it is impossible to change it back through
API. Reason is the check that refuses update operation when there are
more then 1 updated props with immutable switched on. For projects it is
no problem to send 1 update but for domain we pull additional base props
(is_domain=true, domain_id=none, parent_id=none) in the
get_project_from_domain method. And since the test for unsetting
immutable started with immutable already unset (in difference to similar
test for the project) it was never identified.
- discard default props of the domain to keep the length check happy
- ensure test for unsetting immutable starts with immutable domain

Change-Id: I03ba754875050fdb93219e915fc099680679b6c4