Add a new index on project_endpoint_group
The current primary key is a composed key with both `endpoint_group_id`
and `project_id`.
As a result, SQL requests with only `project_id` in WHERE statement
where doing a full scan, slowing down the API and overusing the
database.

Adding a new index on `project_id` is heavily reducing the load on
large-scale environments.

Signed-off-by: Arnaud Morin <arnaud.morin@ovhcloud.com>
Change-Id: I51e6c32f17df9473f9a055013eed1fe0a90c8afa