api: Correct query string schema for access rules API
The presence of a call to 'build_driver_hints' in the method indicates
that filters are permitted for this API, even if they're not currently
documented.

We also rename the variables and fix a docstring to avoid confusion
about what the schemas are intended for.

Change-Id: I17c4e5116ded437a8561b5f721899cdc653c352e
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
(cherry picked from commit 187c1af5068997ab1ebd29cb7deaa8138b496af6)