Separate user response and request schema
There are more then only few attributes only present in the response and
cannot be present in the request. The worst effect is that boolean
(enabled) cannot be returned as all this amount of values supported on
the input. This leads to pretty weird results when using produces
openapi spec with the mock generators.

Copy out user properties adopting them to the reality of responses.

Change-Id: Ic4c81cc8f7b90adb2cb6cea19cf2b3c4bff2c00f
Signed-off-by: Artem Goncharov <artem.goncharov@gmail.com>