Fix AD nested groups issues
The implementation of AD nested groups searches works fine when
listing the groups a user belongs to, but fails when listing all
members of a group. This function of listing all members is also
used to check if a user belongs to a group which also fails.
This patch fixes the query for getting all users in a group.

Closes-Bug: #2112477

Depends-on: https://review.opendev.org/c/openstack/devstack/+/962711
Depends-on: https://review.opendev.org/c/openstack/devstack/+/962704
Depends-on: https://review.opendev.org/c/openstack/keystone/+/962696

Change-Id: I9707e1a9bc4a334902933d6251888144f8c3bc19
Signed-off-by: Jorge Merlino <jorge.merlino@canonical.com>
(cherry picked from commit f8338be43073f23f3db64fa4ba658c3e1f554aa7)