trust schema: don't require project_id to be in uuid format
Not all clouds have uuid project ids; projects might predate
the adoption of uuids in keystone code, or they might draw project
ids from an external provider.

In addition, this constraint is unique to trusts, which results in a
keystone api that works with non-uuid projects 98% of the time but
fails by surprise in this exact case.

For a similar example, see

https: //review.opendev.org/c/openstack/cinder/+/864585
Change-Id: Ic0fe21acb8adfe3030c9b58007456efb4b61799f