Keystone identity mapping to support project definition as a JSON
This pull request (PR) depends on [1], which has been released. The
proposal in [1] enhances the identity mapping schema management.

Currently, the project assignment via the federated identity mapping is
static. This happens because of the find/replace mechanism that we have
in place in Keystone mapping processor. Therefore, if the IdP provider
generates an attribute that contains a JSON with project definitions,
we are not able to handle it in Keystone properly.

This PR introduces a new property in the federated identity mapping
schema called `projects_json` that can also be used as `projects.
In the schema, this property will accept a JSON string, that defines
all of the projects and their specific roles that the user must
receive when login-in to the OpenStack platform. Moreover, when using
this extension, roles (assigned to projects) are added and removed
on the fly.

The extension is quite straight forward. We created a new
``federation_attribute_mapping_schema_version`` version (3.0).
This new version enables the handling of `project_json` and `projects`
in the attribute mapping that can contain a JSON string with the
projects assingment.

Furthermore, we added code to handle the addition of extra roles for
projects and removal of roles that are present in OpenStack, but are
not in the IdP data. This is a mechanism to make the state of the
OpenStack federated user consistent with the Identity provider user
attributes.

[1] https://bugs.launchpad.net/keystone/+bug/1887515

Implements: https://bugs.launchpad.net/keystone/+bug/1888412
Closes-Bug: #1888412

Depends-On: https://review.opendev.org/#/c/739966/

Change-Id: I585a0a15f739620d4f3a7a66e9d4cbc69469da8b
Signed-off-by: Rafael Weingärtner <rafael@apache.org>