Rework `concourse_unit_test_task` (#41)
* tox: do not exclude-regex tests

It doesn't cause failures without the argument.

* Do not clone from remote but use Concourse provided repo

Concourse already has the repo version to be tested cloned under
`source` path.

* Do not run memcached

Executing without memcached doesn't change the number of tests and the
outcome.

* Do not set WATCHER_DISABLED=true

It looks like runtime parameter, not changing tests behavior.

* Make it a normal bash scipt

It doesn't have to be one-line, given the way Concourse runs it.

But working with it as with a regular script is much easier in
development and testing.

* Install tools in `tools` venv

Installing them on Ubunbut 24.04 fails -
https://stackoverflow.com/a/75696359. Whereas an alternative option
could be to use `--break-system-packages` - which is less nicer than
venv IMHO.

* Use latest tox

Upstream tests run with v4+, and our could too.

* Install OS requirements using bindep

The OS requirements packages change over time (some of which are no
longer relevant for e.g. Ubuntu 24.04), and are maintained in
`bindep.txt`. So just install them with `bindep`, similar to how Zuul
does that upstream
- https://opendev.org/zuul/zuul-jobs/src/branch/master/roles/bindep/tasks/packages.yaml.

The disadvantage could be that it could install more than required -
e.g. mysql, postgresql or graphviz, which we don't need for our test
run.

* Add `build-essential` to fix psycopg2, python-ldap

* Install and configure git

Without `safe.directory` it fails to use `git` under `/source`, and
fails to recognize pbr version correspondingly
(https://docs.openstack.org/pbr/latest/user/features.html#version).

* Maintain LDAP tests environment variables

With or without them the number of executed tests and the outcome are
the same. Not sure why so - needs more investigation. But just maintain
them as in the documentation -
https://docs.openstack.org/keystone/latest/contributor/testing-keystone.html#ldap-tests

* Make tox to NOT skip-missing-interpreters

This is to make sure that it won't succeed in case the interpreter could
not be found.