api-ref: Add (deprecated) endpoint name docs
In change I61429b39086d15aed7d2315d7c3971727a9fa419 we
removed documentation for the name attribute since it is not
recommended. However, the api still accepts and (optionally)
returns this field, which can be confusing.

Rather than pretend it does not exist from a docs perspective,
instead indicate that it does exist but should not be used.

Change-Id: I381de2daf422f3328c75b79ed8c8b543290abe23
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>